fix(quick-settings): open 2FA modal from user menu
The Modal lived inside UserMenuQuickSettings, which itself was rendered
inside the {open && (...)} conditional of UserMenu. Clicking "Set up
now" closed the dropdown (setOpen(false)) — which unmounted
UserMenuQuickSettings, the Modal, and its state in the same render
pass, so setSetupOpen(true) was applied to an already-unmounting
component and the Modal never appeared.

Lift the Modal + setupOpen/setupStep state into UserMenu so it lives
outside the conditional dropdown block and survives the close.
